golang-api/grpc/requirements.txt
Flask==3.1.0
flask_cors==5.0.0
Pillow>=10.2.0,<11
svgtrace==2024
svgpathtools==1.6.1
grpcio==1.71.0
grpcio-tools==1.71.0
